refactor: move src/http/auth.ts to src/adapters/http/
